+++ Old UPTS Stuff/Web and Databases/AllPlayers.sql
@@ -1,0 +1,9 @@
+SELECT 
+    `players`.`idplayers`,
+    `players`.`player_name`,
+    `players`.`player_games`,
+    `players`.`player_chars`,
+    `players`.`username`
+FROM
+    `upts_s1`.`players`
+where username = *;
