+++ Old UPTS Stuff/Web and Databases/create_upts.sql
@@ -1,0 +1,203 @@
+-- MySQL Script generated by MySQL Workbench
+-- Thu Nov  8 17:36:00 2018
+-- Model: New Model    Version: 1.0
+-- MySQL Workbench Forward Engineering
+
+SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;
+SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;
+SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';
+
+-- -----------------------------------------------------
+-- Schema upts_s1
+-- -----------------------------------------------------
+
+-- -----------------------------------------------------
+-- Schema upts_s1
+-- -----------------------------------------------------
+CREATE SCHEMA IF NOT EXISTS `upts_s1` DEFAULT CHARACTER SET utf8 ;
+USE `upts_s1` ;
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`games`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`games` (
+  `idgames` INT(11) NOT NULL AUTO_INCREMENT,
+  `game_name` VARCHAR(45) NOT NULL,
+  `game_rel_date` DATETIME NULL,
+  `game_developer` VARCHAR(45) NULL,
+  `game_distributor` VARCHAR(45) NULL,
+  PRIMARY KEY (`idgames`),
+  UNIQUE INDEX `game_name_UNIQUE` (`game_name` ASC) VISIBLE)
+ENGINE = InnoDB
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`achievements`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`achievements` (
+  `idachievements` INT(11) NOT NULL AUTO_INCREMENT,
+  `ach_name` VARCHAR(45) NOT NULL,
+  `ach_desc` VARCHAR(250) NULL DEFAULT NULL,
+  `games_idgames` INT(11) NOT NULL,
+  PRIMARY KEY (`idachievements`),
+  CONSTRAINT `fk_achievements_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`users`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`users` (
+  `idusers` INT(11) NOT NULL AUTO_INCREMENT,
+  `username` VARCHAR(45) NOT NULL,
+  `password` VARCHAR(45) NULL DEFAULT NULL,
+  PRIMARY KEY (`idusers`),
+  UNIQUE INDEX `username_UNIQUE` (`username` ASC) VISIBLE)
+ENGINE = InnoDB
+AUTO_INCREMENT = 8
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`players`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`players` (
+  `idplayers` INT(11) NOT NULL AUTO_INCREMENT,
+  `player_name` VARCHAR(45) NOT NULL,
+  `player_games` VARCHAR(45) NULL DEFAULT NULL,
+  `player_chars` VARCHAR(45) NULL DEFAULT NULL,
+  `users_idusers` INT(11) NOT NULL,
+  PRIMARY KEY (`idplayers`, `player_name`),
+  CONSTRAINT `fk_players_users1`
+    FOREIGN KEY (`users_idusers`)
+    REFERENCES `upts_s1`.`users` (`idusers`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB
+AUTO_INCREMENT = 24
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`trophies`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`trophies` (
+  `idtrophies` INT(11) NOT NULL AUTO_INCREMENT,
+  `trophy_name` VARCHAR(45) NOT NULL,
+  `trophy_description` VARCHAR(250) NULL DEFAULT NULL,
+  `games_idgames` INT(11) NOT NULL,
+  PRIMARY KEY (`idtrophies`),
+  CONSTRAINT `fk_trophies_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`notes`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`notes` (
+  `idnotes` INT(11) NOT NULL AUTO_INCREMENT,
+  `gnote_name` VARCHAR(45) NULL,
+  `gnote_details` VARCHAR(250) NULL,
+  `games_idgames` INT(11) NOT NULL,
+  PRIMARY KEY (`idnotes`),
+  CONSTRAINT `fk_notes_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`currencies`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`currencies` (
+  `idcurrencies` INT NOT NULL AUTO_INCREMENT,
+  `game_currency` VARCHAR(45) NULL,
+  `currency_note` VARCHAR(250) NULL,
+  `games_idgames` INT(11) NOT NULL,
+  PRIMARY KEY (`idcurrencies`),
+  CONSTRAINT `fk_currencies_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`items`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`items` (
+  `iditems` INT NOT NULL AUTO_INCREMENT,
+  `item_name` VARCHAR(45) NOT NULL,
+  `item_desc` VARCHAR(250) NULL,
+  `item_cost` INT NULL,
+  `cost_unit` VARCHAR(45) GENERATED ALWAYS AS (),
+  `games_idgames` INT(11) NOT NULL,
+  PRIMARY KEY (`iditems`),
+  CONSTRAINT `fk_items_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`games_has_users`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`games_has_users` (
+  `games_idgames` INT(11) NOT NULL,
+  `users_idusers` INT(11) NOT NULL,
+  PRIMARY KEY (`games_idgames`, `users_idusers`),
+  CONSTRAINT `fk_games_has_users_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION,
+  CONSTRAINT `fk_games_has_users_users1`
+    FOREIGN KEY (`users_idusers`)
+    REFERENCES `upts_s1`.`users` (`idusers`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB
+DEFAULT CHARACTER SET = utf8;
+
+
+-- -----------------------------------------------------
+-- Table `upts_s1`.`games_has_players`
+-- -----------------------------------------------------
+CREATE TABLE IF NOT EXISTS `upts_s1`.`games_has_players` (
+  `games_idgames` INT(11) NOT NULL,
+  `players_idplayers` INT(11) NOT NULL,
+  `players_player_name` VARCHAR(45) NOT NULL,
+  PRIMARY KEY (`games_idgames`, `players_idplayers`, `players_player_name`),
+  CONSTRAINT `fk_games_has_players_games1`
+    FOREIGN KEY (`games_idgames`)
+    REFERENCES `upts_s1`.`games` (`idgames`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION,
+  CONSTRAINT `fk_games_has_players_players1`
+    FOREIGN KEY (`players_idplayers` , `players_player_name`)
+    REFERENCES `upts_s1`.`players` (`idplayers` , `player_name`)
+    ON DELETE NO ACTION
+    ON UPDATE NO ACTION)
+ENGINE = InnoDB
+DEFAULT CHARACTER SET = utf8;
+
+
+SET SQL_MODE=@OLD_SQL_MODE;
+SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
+SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
