+++ Old UPTS Stuff/Web and Databases/minitest.html
@@ -1,0 +1,11 @@
+
+<!DOCTYPE html>
+<html>
+<body>
+
+<?php
+echo "My first PHP script!";
+?>
+
+</body>
+</html>
