+++ Old UPTS Stuff/Web and Databases/php-to-sql-test.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+<body>
+
+<?php
+$servername = "127.0.0.1";
+$username = "testuser";
+$password = "testpass";
+
+// Create connection
+$conn = new mysqli($servername, $username, $password);
+
+// Check connection
+if ($conn->connect_error) {
+    die("Connection failed: " . $conn->connect_error);
+}
+echo "Connected successfully";
+?> 
+
+
+</body>
+</html>
