+++ Old UPTS Stuff/Web and Databases/upts/polls/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class PollsConfig(AppConfig):
+    name = 'polls'
