+++ Old UPTS Stuff/Web and Databases/upts/polls/tests.py
@@ -1,0 +1,3 @@
+from django.test import TestCase
+
+# Create your tests here.
