+++ Old UPTS Stuff/Web and Databases/upts/polls/views.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+from django.http import HttpResponse
+
+# Create your views here.
+
+def index(request):
+    return HttpResponse("Hello, world. You're at the polls index.")
