+++ README.txt
@@ -11,74 +11,6 @@ User Stories:
 
 User = dev, designer, engineer or QA
 
-1)	Login
-	⁃	As a user I need to be able to log in so I can access the features of the system.  
-	⁃	As a user I need to be able to retrieve lost login info or create a new user account so that I can access the features of the system.
-	⁃	Description: Registration screen for the developer to access the system
-	⁃	Description Login screen for a registered user
-	⁃	Description Offer to help the user retrieve their info
 
-2)	Player Tracking and Reporting
-	⁃	As a user I want to be able to add multiple players to a single account so I can get aggregate data easily.
-	⁃	Description Form for developer to add a player’s information
-	⁃	Description: Form for developer to modify a player’s information
-	⁃	Player Report
-	⁃	Description: Screen for developer to request Player Report
-
-3)	Game Tracking and Reporting
-	⁃	As a user I want to be able to Add multiple games to a given player so I can get aggregate data easily.
-	⁃	As a user I want to be able to easily access usual data about each game like inventory, currency or trophies.
-	⁃	As a user I want to be able to add custom achievements so I can track specific gameplay elements that may not be otherwise defined.
-	⁃	As a user I want to be able to add multiple characters to a single game so I can track aggregate data.
-	⁃	Description: Form for developer to Add a Game and its Features
-	⁃	Description: Form to add characters to a player
-	⁃	Game Report
-	⁃	Description: Screen for developer to request Game Report
-	⁃	Items Report
-	⁃	Description: Sub Screen below Player, Character, or Game report for developer to request Items Report
-
-
-Additional Agile Development Docs:
-
-All the documents used in the Agile Development of this project - including out of date documents that may have changed over the course of the project.
-
-https://github.com/WillWagnerIV/The-Coconuts/tree/master/Docs/Agile%20Docs
-
-Additional Program Documentation:
-
-Additional documents that explain the use of the program rather than the development process.  Because this is agile - there will be few if any documents in this folder.
-
-https://github.com/WillWagnerIV/The-Coconuts/tree/master/Docs/Application%20Docs
-
-Milestone 1 - Nov. 8 2018
-
-	•	Completed setup of remote database server
-	•	Completed Basic Navigation menus in Python application
-	•	Completed successful connections to the remote databases from the python application
-	•	Completed test functions for read / write / create / delete databases from the python application
-	•	Completed test functions and established a format fo read / write to json files.
-
-Backlog:
-	•	Implement functionality for remaining few existing menu options that are not completed.
-
-Milestone 2 - Dec. 6 2018
-
-1)	Complete remaining milestone 1.0 backlog.
-
-2)	UI
-	⁃	Django
-	⁃	Insights
-
-3)	Implement testing 
-	⁃	Usability testing 
-	⁃	Menus Organization 
-	⁃	Ease of use
-4)	pytest Functions
-	⁃	Add/Delete/Modify User
-	⁃	Add/Delete/Modify Game
-	⁃	Add/Delete/Modify Players
-	⁃	2 Sample Reports
-	⁃	Test Database Read / Write
-	⁃	Test json Read / Write
