+++ UPTS/json/Symbol Game.json
@@ -1,0 +1,6 @@
+{"Symbol Game":{"game_name":"Symbol Game"}}
+{"Symbol Game":{"game_notes":[{"Note 1":"A matching game with symbols."},{"Note 2":"Score is proportional with the number of correct answers"}]}}
+{"Symbol Game":{"game_currency":[{"None":"N\/A"}]}}
+{"Symbol Game":{"game_trophies":[{"None":"N\/A"}]}}
+{"Symbol Game":{"game_ach":[{"Completed Level One on Normal Difficulty":"40%"}]}}, 
+{"Symbol Game":{"game_items":[{"None":"N\/A"}]}}
